backend_part/core/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Django==5.1.2
et-xmlfile==1.1.0
executing==2.1.0
faiss-cpu==1.9.0
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jedi==0.19.2
Jinja2==3.1.6
jiter==0.7.1
jupyter_core==5.7.2
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.68.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.26.5
wcwidth==0.2.13
whisper==1.1.10
